requirements.txt
streamlit>=1.34.0
gspread>=6.0.0
oauth2client>=4.1.3
pandas>=2.0.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
pywhatkit>=5.4
google-genai>=1.0.0
